part3/hbnb/requirements.txt
flask
flask-restx
flask-bcrypt
flask_cors
flask-jwt-extended
sqlalchemy
flask-SQLAlchemy
